requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
mysql-connector-python==8.2.0
python-dotenv==1.0.0
aiohttp==3.9.1
passlib==1.7.4
python-multipart==0.0.6
jinja2==3.1.2
requests==2.31.0

# ML Dependencies
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.24.3
joblib==1.3.2
beautifulsoup4==4.12.2
reportlab==4.2.2
Pillow==10.3.0